% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSATtools-package.R
\docType{package}
\name{RSATtools-package}
\alias{RSATtools}
\alias{RSATtools-package}
\title{RSATtools: RSATtools: an R package interacting with the RSAT web services}
\description{
This package provides an interface to the RSAT collection of
    web services http://embnet.ccg.unam.mx/rsat/ from R and is integrated in 
    the Bioconductor ecosystem. This package was created
    as a collaboration by members of the Community of Bioinformatics
    Software Developers (CDSB in Spanish) https://comunidadbioinfo.github.io/.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ComunidadBioInfo/RSATtools}
  \item Report bugs at \url{https://github.com/ComunidadBioInfo/RSATtools/issues}
}

}
\author{
\strong{Maintainer}: Joselyn Chavez \email{joselynchavezf@gmail.com} (\href{https://orcid.org/0000-0002-4974-4591}{ORCID})

Authors:
\itemize{
  \item Carmina Barberena-Jonas \email{car.barjon@gmail.com}
  \item Jesus E. Sotelo-Fonseca \email{jemilianosf@gmail.com}
}

Other contributors:
\itemize{
  \item Jose Alquicira-Hernandez \email{joseah@lcg.unam.mx} (\href{https://orcid.org/0000-0002-9049-7780}{ORCID}) [contributor]
  \item Heladia Salgado \email{heladia@ccg.unam.mx} (\href{https://orcid.org/0000-0002-3166-5801}{ORCID}) [contributor]
  \item Leonardo Collado-Torres \email{lcolladotor@gmail.com} (\href{https://orcid.org/0000-0003-2140-308X}{ORCID}) [contributor]
  \item Alejandro Reyes \email{alejandro.reyes.ds@gmail.com} (\href{https://orcid.org/0000-0001-8717-6612}{ORCID}) [contributor]
}

}
\keyword{internal}
